+++ src/lib.rs
@@ -51,229 +51,11 @@
 )]
 #![deny(clippy::all, clippy::pedantic)]
 
-use {
-    core::{
-        marker::PhantomData,
-        ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Sub, SubAssign},
-    },
-    x86_64::structures::paging::PageSize,
-};
+mod bytes;
+mod num_of_pages;
 
-#[repr(transparent)]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-/// A struct representing byte size.
-pub struct Bytes(usize);
-impl Bytes {
-    /// Creates a new instance with given value.
-    #[must_use]
-    pub const fn new(bytes: usize) -> Self {
-        Self(bytes)
-    }
-
-    /// Equivalent to `Bytes::new(0)`.
-    #[must_use]
-    pub const fn zero() -> Self {
-        Self::new(0)
-    }
-
-    /// Returns the value.
-    #[must_use]
-    pub const fn as_usize(self) -> usize {
-        self.0
-    }
-
-    /// Converts bytes to the number of physical pages. Note that the number of physical pages will
-    /// be calculated so that the specified bytes will be fit in pages.
-    #[must_use]
-    pub const fn as_num_of_pages<T: PageSize>(self) -> NumOfPages<T> {
-        #[allow(clippy::cast_possible_truncation)]
-        NumOfPages::new((self.0 + T::SIZE as usize - 1) / T::SIZE as usize)
-    }
-}
-impl Add for Bytes {
-    type Output = Bytes;
-
-    fn add(self, rhs: Bytes) -> Self {
-        Self::new(self.0 + rhs.0)
-    }
-}
-impl Add<usize> for Bytes {
-    type Output = Bytes;
-
-    fn add(self, rhs: usize) -> Self::Output {
-        Self::new(self.0 + rhs)
-    }
-}
-impl AddAssign for Bytes {
-    fn add_assign(&mut self, rhs: Bytes) {
-        self.0 += rhs.0;
-    }
-}
-impl AddAssign<usize> for Bytes {
-    fn add_assign(&mut self, rhs: usize) {
-        self.0 += rhs;
-    }
-}
-impl Sub for Bytes {
-    type Output = Bytes;
-
-    fn sub(self, rhs: Bytes) -> Self {
-        Self::new(self.0 - rhs.0)
-    }
-}
-impl Sub<usize> for Bytes {
-    type Output = Bytes;
-
-    fn sub(self, rhs: usize) -> Self::Output {
-        Self::new(self.0 - rhs)
-    }
-}
-impl SubAssign for Bytes {
-    fn sub_assign(&mut self, rhs: Bytes) {
-        self.0 -= rhs.0;
-    }
-}
-impl SubAssign<usize> for Bytes {
-    fn sub_assign(&mut self, rhs: usize) {
-        *self -= Bytes::new(rhs);
-    }
-}
-impl Mul<usize> for Bytes {
-    type Output = Bytes;
-    fn mul(self, rhs: usize) -> Self::Output {
-        Self(self.0 * rhs)
-    }
-}
-impl MulAssign<usize> for Bytes {
-    fn mul_assign(&mut self, rhs: usize) {
-        *self = *self * rhs;
-    }
-}
-impl Div<usize> for Bytes {
-    type Output = Bytes;
-
-    fn div(self, rhs: usize) -> Self::Output {
-        Self(self.0 / rhs)
-    }
-}
-impl DivAssign<usize> for Bytes {
-    fn div_assign(&mut self, rhs: usize) {
-        *self = *self / rhs;
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-/// A struct representing the number of pages.
-pub struct NumOfPages<T: PageSize> {
-    num_of_pages: usize,
-    _marker: PhantomData<fn() -> T>,
-}
-impl<T: PageSize> NumOfPages<T> {
-    /// Creates a new instance with given value.
-    #[must_use]
-    pub const fn new(num_of_pages: usize) -> Self {
-        Self {
-            num_of_pages,
-            _marker: PhantomData,
-        }
-    }
-
-    /// Equivalent to `NumOfPages::new(0)`.
-    #[must_use]
-    pub const fn zero() -> Self {
-        Self::new(0)
-    }
-
-    /// Returns the value.
-    #[must_use]
-    pub const fn as_usize(self) -> usize {
-        self.num_of_pages
-    }
-
-    /// Converts the number of physical pages to bytes.
-    #[must_use]
-    pub const fn as_bytes(self) -> Bytes {
-        #[allow(clippy::cast_possible_truncation)]
-        Bytes::new(self.num_of_pages * T::SIZE as usize)
-    }
-}
-impl<T: PageSize> Add for NumOfPages<T> {
-    type Output = NumOfPages<T>;
-
-    fn add(self, rhs: NumOfPages<T>) -> Self {
-        Self::new(self.num_of_pages + rhs.num_of_pages)
-    }
-}
-impl<T: PageSize> Add<usize> for NumOfPages<T> {
-    type Output = NumOfPages<T>;
-
-    fn add(self, rhs: usize) -> Self::Output {
-        Self::new(self.num_of_pages + rhs)
-    }
-}
-impl<T: PageSize> AddAssign for NumOfPages<T> {
-    fn add_assign(&mut self, rhs: NumOfPages<T>) {
-        self.num_of_pages += rhs.num_of_pages;
-    }
-}
-impl<T: PageSize> AddAssign<usize> for NumOfPages<T> {
-    fn add_assign(&mut self, rhs: usize) {
-        self.num_of_pages += rhs;
-    }
-}
-impl<T: PageSize> Sub for NumOfPages<T> {
-    type Output = NumOfPages<T>;
-
-    fn sub(self, rhs: NumOfPages<T>) -> Self {
-        Self::new(self.num_of_pages - rhs.num_of_pages)
-    }
-}
-impl<T: PageSize> Sub<usize> for NumOfPages<T> {
-    type Output = NumOfPages<T>;
-
-    fn sub(self, rhs: usize) -> Self::Output {
-        Self::new(self.num_of_pages - rhs)
-    }
-}
-impl<T: PageSize> SubAssign for NumOfPages<T> {
-    fn sub_assign(&mut self, rhs: NumOfPages<T>) {
-        self.num_of_pages -= rhs.num_of_pages;
-    }
-}
-impl<T: PageSize> SubAssign<usize> for NumOfPages<T> {
-    fn sub_assign(&mut self, rhs: usize) {
-        *self -= Self::new(rhs);
-    }
-}
-impl<T: PageSize> Mul<usize> for NumOfPages<T> {
-    type Output = NumOfPages<T>;
-    fn mul(self, rhs: usize) -> Self::Output {
-        Self {
-            num_of_pages: self.num_of_pages * rhs,
-            ..self
-        }
-    }
-}
-impl<T: PageSize> MulAssign<usize> for NumOfPages<T> {
-    fn mul_assign(&mut self, rhs: usize) {
-        *self = *self * rhs;
-    }
-}
-impl<T: PageSize> Div<usize> for NumOfPages<T> {
-    type Output = NumOfPages<T>;
-
-    fn div(self, rhs: usize) -> Self::Output {
-        Self {
-            num_of_pages: self.num_of_pages / rhs,
-            ..self
-        }
-    }
-}
-impl<T: PageSize> DivAssign<usize> for NumOfPages<T> {
-    fn div_assign(&mut self, rhs: usize) {
-        *self = *self / rhs;
-    }
-}
+pub use bytes::Bytes;
+pub use num_of_pages::NumOfPages;
 
 #[cfg(test)]
 mod tests {
